setup tasks independent of AnalysisAlien plugin
I think this is the right way to go, but some analysis seem to use dark
magic that breaks the streaming(?) to the grid. Current solution works
with simple tasks, eg:

\#_______________________________________________________________________________
\#Module.Begin        PhysicsSelection
\#Module.Libs
\#Module.DataTypes    ESD, AOD, MC
\#Module.MacroName    $ALICE_PHYSICS/OADB/macros/AddTaskPhysicsSelection.C
\#Module.MacroArgs    kFALSE, kTRUE
\#Module.Deps
\#Module.Owner        cjahnke
\#Module.OutputFile   jpsi_Default.root
\#Module.StartConfig

\#Module.EndConfig

This works in proof lite as well as on the grid.

The advantage of loading the task this way is that it is identical for
running locally and on the grid and thus reduces possible errors down
the road